[refactor] 인스펙터창 분리하기
