+++ Engine/src/Editor/4.txt
@@ -1,0 +1,40 @@
+폴더 PATH의 목록입니다.
+볼륨 일련 번호는 285E-FFD7입니다.
+D:.
+│  4.txt
+│  
+├─AssetEditors
+│      MaterialAssetEditor.cpp
+│      UICurveAssetEditor.cpp
+│      
+├─Core
+│      EditorCommands.h
+│      EditorCore.cpp
+│      EditorCore.h
+│      EditorUIState.h
+│      EditorUndoRedo.h
+│      ReflectionUI.h
+│      ViewportPicker.cpp
+│      ViewportPicker.h
+│      
+├─Panels
+│      CameraPanel.cpp
+│      GameViewportPanel.cpp
+│      HierarchyPanel.cpp
+│      InspectorPanel.cpp
+│      LightingPanel.cpp
+│      ProjectPanel.cpp
+│      
+├─Tools
+│  │  BuildGameWindow.cpp
+│  │  PvdSettingsWindow.cpp
+│  │  
+│  └─Blueprint
+│          AnimBlueprintEditor.cpp
+│          AnimBlueprintEditor.h
+│          
+└─UI
+        EditorDocking.cpp
+        EditorMainMenuBar.cpp
+        EditorSceneLoadFlow.cpp
+
