+++ Engine/src/Editor/Inspector/Inspector_Camera.cpp
@@ -1,0 +1,127 @@
+#include "Editor/Core/EditorCore.h"
+#include "Editor/Core/EditorUIState.h"
+
+namespace Alice
+{
+void EditorCore::DrawInspectorCameraSpringArm(World& world, const EntityId& _selectedEntity)
+	{
+		if (auto* comp = world.GetComponent<CameraSpringArmComponent>(_selectedEntity))
+		{
+			if (ImGui::CollapsingHeader("Camera Spring Arm", ImGuiTreeNodeFlags_DefaultOpen))
+			{
+				bool changed = false;
+				changed |= ReflectionUI::RenderProperty(*comp, "enabled", "Enabled", &world).changed;
+				changed |= ReflectionUI::RenderProperty(*comp, "enableCollision", "Enable Collision", &world).changed;
+				changed |= ReflectionUI::RenderProperty(*comp, "enableZoom", "Enable Zoom", &world).changed;
+				changed |= ReflectionUI::RenderProperty(*comp, "distance", "Distance", &world).changed;
+				changed |= ReflectionUI::RenderProperty(*comp, "minDistance", "Min Distance", &world).changed;
+				changed |= ReflectionUI::RenderProperty(*comp, "maxDistance", "Max Distance", &world).changed;
+				changed |= ReflectionUI::RenderProperty(*comp, "zoomSpeed", "Zoom Speed", &world).changed;
+				changed |= ReflectionUI::RenderProperty(*comp, "distanceDamping", "Distance Damping", &world).changed;
+				changed |= ReflectionUI::RenderProperty(*comp, "probeRadius", "Probe Radius", &world).changed;
+				changed |= ReflectionUI::RenderProperty(*comp, "probePadding", "Probe Padding", &world).changed;
+				changed |= ReflectionUI::RenderProperty(*comp, "minHeight", "Min Height", &world).changed;
+				
+				if (changed) g_SceneDirty = true;
+			}
+		}
+	}
+
+
+void EditorCore::DrawInspectorCameraLookAt(World& world, const EntityId& _selectedEntity)
+	{
+		if (auto* comp = world.GetComponent<CameraLookAtComponent>(_selectedEntity))
+		{
+			if (ImGui::CollapsingHeader("Camera Look At", ImGuiTreeNodeFlags_DefaultOpen))
+			{
+				bool changed = false;
+				changed |= ReflectionUI::RenderProperty(*comp, "enabled", "Enabled", &world).changed;
+				changed |= ReflectionUI::RenderProperty(*comp, "targetName", "Target Name", &world).changed;
+				changed |= ReflectionUI::RenderProperty(*comp, "rotationDamping", "Rotation Damping", &world).changed;
+				
+				if (changed) g_SceneDirty = true;
+			}
+		}
+	}
+
+
+void EditorCore::DrawInspectorCameraFollow(World& world, const EntityId& _selectedEntity)
+	{
+		if (auto* comp = world.GetComponent<CameraFollowComponent>(_selectedEntity))
+		{
+			if (ImGui::CollapsingHeader("Camera Follow", ImGuiTreeNodeFlags_DefaultOpen))
+			{
+				bool changed = false;
+				
+				// 런타임 상태 필터 (직렬화/편집 대상 아님)
+				auto filter = [](const std::string& propName) -> bool {
+					// 런타임 상태 필드들은 인스펙터에서 제외
+					return propName != "lockOnTargetId" && 
+					       propName != "lockOnActive" && 
+					       propName != "initialized" &&
+					       propName != "smoothedPosition" &&
+					       propName != "smoothedRotation";
+				};
+				
+				changed |= ReflectionUI::RenderInspector(*comp, filter, &world).changed;
+				
+				if (changed) g_SceneDirty = true;
+			}
+		}
+	}
+
+
+void EditorCore::DrawInspectorCameraShake(World& world, const EntityId& _selectedEntity)
+	{
+		if (auto* comp = world.GetComponent<CameraShakeComponent>(_selectedEntity))
+		{
+			if (ImGui::CollapsingHeader("Camera Shake", ImGuiTreeNodeFlags_DefaultOpen))
+			{
+				bool changed = false;
+				changed |= ReflectionUI::RenderProperty(*comp, "enabled", "Enabled", &world).changed;
+				changed |= ReflectionUI::RenderProperty(*comp, "amplitude", "Amplitude", &world).changed;
+				changed |= ReflectionUI::RenderProperty(*comp, "frequency", "Frequency", &world).changed;
+				changed |= ReflectionUI::RenderProperty(*comp, "duration", "Duration", &world).changed;
+				changed |= ReflectionUI::RenderProperty(*comp, "decay", "Decay", &world).changed;
+				
+				if (changed) g_SceneDirty = true;
+			}
+		}
+	}
+
+
+void EditorCore::DrawInspectorCameraInput(World& world, const EntityId& _selectedEntity)
+	{
+		if (auto* comp = world.GetComponent<CameraInputComponent>(_selectedEntity))
+		{
+			if (ImGui::CollapsingHeader("Camera Input", ImGuiTreeNodeFlags_DefaultOpen))
+			{
+				bool changed = false;
+				changed |= ReflectionUI::RenderInspector(*comp, nullptr, &world).changed;
+				
+				if (changed) g_SceneDirty = true;
+			}
+		}
+	}
+
+
+void EditorCore::DrawInspectorCameraBlend(World& world, const EntityId& _selectedEntity)
+	{
+		if (auto* comp = world.GetComponent<CameraBlendComponent>(_selectedEntity))
+		{
+			if (ImGui::CollapsingHeader("Camera Blend", ImGuiTreeNodeFlags_DefaultOpen))
+			{
+				bool changed = false;
+				changed |= ReflectionUI::RenderProperty(*comp, "active", "Active", &world).changed;
+				changed |= ReflectionUI::RenderProperty(*comp, "targetName", "Target Name", &world).changed;
+				changed |= ReflectionUI::RenderProperty(*comp, "duration", "Duration", &world).changed;
+				changed |= ReflectionUI::RenderProperty(*comp, "useSmoothStep", "Use Smooth Step", &world).changed;
+				changed |= ReflectionUI::RenderProperty(*comp, "slowTriggerT", "Slow Trigger T", &world).changed;
+				changed |= ReflectionUI::RenderProperty(*comp, "slowDuration", "Slow Duration", &world).changed;
+				changed |= ReflectionUI::RenderProperty(*comp, "slowTimeScale", "Slow Time Scale", &world).changed;
+				
+				if (changed) g_SceneDirty = true;
+			}
+		}
+	}
+}
