+++ Engine/src/Editor/Inspector/Inspector_Physics.cpp
@@ -1,0 +1,1480 @@
+#include "Editor/Core/EditorCore.h"
+#include "Editor/Core/EditorUIState.h"
+#include "Runtime/Physics/Components/Phy_SettingsComponent.h"
+#include "Runtime/Gameplay/Combat/CombatPhysicsLayers.h"
+#include "Runtime/ECS/GameObject.h"
+#include <array>
+#include <set>
+
+namespace Alice
+{
+bool EditorCore::DrawLayerMaskEditor(const char* label, uint32_t& mask, const std::array<std::string, 32>& layerNames)
+	{
+		bool changed = false;
+		ImGui::Text("%s", label);
+		ImGui::Indent();
+
+		// 최대 32개 레이어를 2열로 표시
+		for (int i = 0; i < 32; i++)
+		{
+			bool bit = (mask & (1u << i)) != 0;
+			std::string layerName = layerNames[i].empty() ? ("Layer " + std::to_string(i)) : layerNames[i];
+			std::string checkboxLabel = layerName + "##" + label + std::to_string(i);
+
+			if (ImGui::Checkbox(checkboxLabel.c_str(), &bit))
+			{
+				if (bit)
+					mask |= (1u << i);
+				else
+					mask &= ~(1u << i);
+				changed = true;
+			}
+
+			// 2열로 배치
+			if ((i + 1) % 2 == 0)
+				ImGui::SameLine();
+		}
+
+		ImGui::Unindent();
+		return changed;
+	}
+
+
+bool EditorCore::DrawLayerMaskChipEditor(const char* label, uint32_t& mask, const std::array<std::string, 32>& layerNames)
+	{
+		bool changed = false;
+		ImGui::Text("%s", label);
+		ImGui::Indent();
+
+		// 현재 선택된 레이어들을 칩으로 표시
+		bool hasAnyLayers = false;
+		for (int i = 0; i < 32; ++i)
+		{
+			if ((mask & (1u << i)) != 0)
+			{
+				hasAnyLayers = true;
+
+				// 레이어 이름
+				std::string layerName = layerNames[i].empty() ? ("Layer " + std::to_string(i)) : layerNames[i];
+
+				// 칩 스타일 버튼 (레이어 이름) - 클릭하면 토글
+				ImGui::PushStyleColor(ImGuiCol_Button, ImVec4(0.4f, 0.6f, 0.9f, 1.0f));
+				ImGui::PushStyleColor(ImGuiCol_ButtonHovered, ImVec4(0.5f, 0.7f, 1.0f, 1.0f));
+				ImGui::PushStyleColor(ImGuiCol_ButtonActive, ImVec4(0.3f, 0.5f, 0.8f, 1.0f));
+
+				std::string chipLabel = layerName + "##" + label + "_chip_" + std::to_string(i);
+				if (ImGui::Button(chipLabel.c_str()))
+				{
+					mask &= ~(1u << i); // 토글: 제거
+					changed = true;
+				}
+
+				ImGui::PopStyleColor(3);
+
+				// 다음 줄로 넘어가기 위해
+				ImGui::SameLine(0.0f, 4.0f);
+			}
+		}
+
+		// 줄바꿈이 필요하면
+		if (hasAnyLayers)
+		{
+			ImGui::NewLine();
+		}
+
+		// + 버튼 (레이어 추가)
+		ImGui::PushStyleColor(ImGuiCol_Button, ImVec4(0.2f, 0.6f, 0.2f, 1.0f));
+		ImGui::PushStyleColor(ImGuiCol_ButtonHovered, ImVec4(0.3f, 0.7f, 0.3f, 1.0f));
+		ImGui::PushStyleColor(ImGuiCol_ButtonActive, ImVec4(0.1f, 0.5f, 0.1f, 1.0f));
+
+		std::string addButtonLabel = "+##" + std::string(label) + "_add";
+		if (ImGui::SmallButton(addButtonLabel.c_str()))
+		{
+			ImGui::OpenPopup((std::string("AddLayer##") + label).c_str());
+		}
+
+		ImGui::PopStyleColor(3);
+
+		// 팝업: 레이어 선택 (선택되지 않은 레이어만 표시)
+		if (ImGui::BeginPopup((std::string("AddLayer##") + label).c_str()))
+		{
+			ImGui::Text("Add Layer");
+			ImGui::Separator();
+
+			bool foundAny = false;
+			for (int i = 0; i < 32; ++i)
+			{
+				if ((mask & (1u << i)) == 0) // 선택되지 않은 레이어만 표시
+				{
+					foundAny = true;
+					std::string layerName = layerNames[i].empty() ? ("Layer " + std::to_string(i)) : layerNames[i];
+					if (ImGui::Selectable(layerName.c_str()))
+					{
+						mask |= (1u << i);
+						changed = true;
+						ImGui::CloseCurrentPopup();
+					}
+				}
+			}
+
+			if (!foundAny)
+			{
+				ImGui::TextDisabled("All layers are selected");
+			}
+
+			ImGui::EndPopup();
+		}
+
+		ImGui::Unindent();
+		return changed;
+	}
+
+
+bool EditorCore::DrawIgnoreLayersChipEditor(const char* label, uint32_t& ignoreLayers, const std::array<std::string, 32>& layerNames)
+	{
+		bool changed = false;
+		ImGui::Text("%s", label);
+		ImGui::Indent();
+
+		// 현재 선택된 레이어들을 칩으로 표시
+		bool hasAnyLayers = false;
+		for (int i = 0; i < 32; ++i)
+		{
+			if ((ignoreLayers & (1u << i)) != 0)
+			{
+				hasAnyLayers = true;
+
+				// 레이어 이름
+				std::string layerName = layerNames[i].empty() ? ("Layer " + std::to_string(i)) : layerNames[i];
+
+				// 칩 스타일 버튼 (레이어 이름) - 클릭해도 아무 일도 안 일어남 (표시만)
+				ImGui::PushStyleColor(ImGuiCol_Button, ImVec4(0.3f, 0.5f, 0.8f, 1.0f));
+				ImGui::PushStyleColor(ImGuiCol_ButtonHovered, ImVec4(0.4f, 0.6f, 0.9f, 1.0f));
+				ImGui::PushStyleColor(ImGuiCol_ButtonActive, ImVec4(0.3f, 0.5f, 0.8f, 1.0f));
+
+				std::string chipLabel = layerName + "##" + label + "_chip_" + std::to_string(i);
+				ImGui::Button(chipLabel.c_str()); // 버튼으로 표시만 (클릭 비활성화)
+
+				ImGui::PopStyleColor(3);
+
+				ImGui::SameLine(0.0f, 4.0f);
+
+				// [x] 버튼 (제거용)
+				ImGui::PushStyleColor(ImGuiCol_Button, ImVec4(0.8f, 0.3f, 0.3f, 1.0f));
+				ImGui::PushStyleColor(ImGuiCol_ButtonHovered, ImVec4(0.9f, 0.4f, 0.4f, 1.0f));
+				ImGui::PushStyleColor(ImGuiCol_ButtonActive, ImVec4(0.7f, 0.2f, 0.2f, 1.0f));
+
+				std::string removeLabel = std::string(" [x]##") + label + "_remove_" + std::to_string(i);
+				if (ImGui::SmallButton(removeLabel.c_str()))
+				{
+					ignoreLayers &= ~(1u << i);
+					changed = true;
+				}
+
+				ImGui::PopStyleColor(3);
+
+				// 다음 줄로 넘어가기 위해
+				ImGui::SameLine(0.0f, 0.0f);
+			}
+		}
+
+		// 줄바꿈이 필요하면
+		if (hasAnyLayers)
+		{
+			ImGui::NewLine();
+		}
+
+		// + 버튼 (레이어 추가)
+		ImGui::PushStyleColor(ImGuiCol_Button, ImVec4(0.2f, 0.6f, 0.2f, 1.0f));
+		ImGui::PushStyleColor(ImGuiCol_ButtonHovered, ImVec4(0.3f, 0.7f, 0.3f, 1.0f));
+		ImGui::PushStyleColor(ImGuiCol_ButtonActive, ImVec4(0.1f, 0.5f, 0.1f, 1.0f));
+
+		std::string addButtonLabel = "+##" + std::string(label) + "_add";
+		if (ImGui::SmallButton(addButtonLabel.c_str()))
+		{
+			ImGui::OpenPopup((std::string("AddIgnoreLayer##") + label).c_str());
+		}
+
+		ImGui::PopStyleColor(3);
+
+		// 팝업: 레이어 선택
+		if (ImGui::BeginPopup((std::string("AddIgnoreLayer##") + label).c_str()))
+		{
+			ImGui::Text("Select layer to ignore:");
+			ImGui::Separator();
+
+			for (int i = 0; i < 32; ++i)
+			{
+				// 이미 추가된 레이어는 표시하지 않음
+				if ((ignoreLayers & (1u << i)) != 0)
+					continue;
+
+				std::string layerName = layerNames[i].empty() ? ("Layer " + std::to_string(i)) : layerNames[i];
+				std::string selectLabel = layerName + "##" + label + "_select_" + std::to_string(i);
+
+				if (ImGui::Selectable(selectLabel.c_str()))
+				{
+					ignoreLayers |= (1u << i);
+					changed = true;
+					ImGui::CloseCurrentPopup();
+				}
+			}
+
+			ImGui::EndPopup();
+		}
+
+		ImGui::Unindent();
+		return changed;
+	}
+
+
+void EditorCore::DrawInspectorCollider(World& world, const EntityId& _selectedEntity)
+	{
+		if (auto* collider = world.GetComponent<Phy_ColliderComponent>(_selectedEntity))
+		{
+			if (ImGui::CollapsingHeader("Collider", ImGuiTreeNodeFlags_DefaultOpen))
+			{
+				bool changed = false;
+
+				if (ImGui::Button("Remove##ColliderRemove"))
+				{
+					world.RemoveComponent<Phy_ColliderComponent>(_selectedEntity);
+					g_SceneDirty = true;
+					return;
+				}
+
+				changed |= ImGui::Checkbox("Debug Draw", &collider->debugDraw);
+
+				// Collider Type 선택
+				ImGui::Text("Collider Type");
+				ImGui::Indent();
+				{
+					const char* typeLabels[] = { "Box", "Sphere", "Capsule" };
+					int typeIndex = static_cast<int>(collider->type);
+					if (ImGui::Combo("##ColliderType", &typeIndex, typeLabels, IM_ARRAYSIZE(typeLabels)))
+					{
+						collider->type = static_cast<ColliderType>(typeIndex);
+						changed = true;
+					}
+				}
+				ImGui::Unindent();
+
+				// 기본 프로퍼티는 ReflectionUI로
+				// Collider 편집 이벤트 처리
+				static EntityId lastEditedColliderEntity = InvalidEntityId;
+				static JsonRttr::json colliderEditStartJson;
+
+				ReflectionUI::UIEditEvent event = ReflectionUI::RenderInspector(*collider, [](const std::string& name) {
+					// type, layerBits는 커스텀 UI로 처리 (collideMask/queryMask는 레이어 매트릭스로만 결정)
+					return name != "type" && name != "layerBits" && name != "physicsActorHandle";
+				});
+
+				// 편집 시작
+				if (event.activated && _selectedEntity != lastEditedColliderEntity)
+				{
+					rttr::instance inst = *collider;
+					colliderEditStartJson = JsonRttr::ToJsonObject(inst);
+					lastEditedColliderEntity = _selectedEntity;
+				}
+
+				// 편집 종료
+				if (event.deactivatedAfterEdit && _selectedEntity == lastEditedColliderEntity)
+				{
+					rttr::instance inst = *collider;
+					JsonRttr::json colliderEditEndJson = JsonRttr::ToJsonObject(inst);
+
+					if (colliderEditStartJson != colliderEditEndJson)
+					{
+						Phy_ColliderComponent oldCollider, newCollider;
+						rttr::instance oldInst = oldCollider;
+						rttr::instance newInst = newCollider;
+						JsonRttr::FromJsonObject(oldInst, colliderEditStartJson);
+						JsonRttr::FromJsonObject(newInst, colliderEditEndJson);
+						PushCommand(std::make_unique<ComponentEditCommand<Phy_ColliderComponent>>(_selectedEntity, oldCollider, newCollider));
+						g_SceneDirty = true;
+					}
+
+					lastEditedColliderEntity = InvalidEntityId;
+				}
+
+				changed |= event.changed;
+
+				// 레이어 마스크 편집
+				ImGui::Separator();
+				ImGui::Text("Layer Settings");
+
+				// Phy_SettingsComponent에서 레이어 이름 가져오기
+				std::array<std::string, 32> layerNames;
+				for (int i = 0; i < 32; ++i)
+					layerNames[i] = "Layer " + std::to_string(i);
+
+				const auto& settingsMap = world.GetComponents<Phy_SettingsComponent>();
+				if (!settingsMap.empty())
+				{
+					const auto& settings = settingsMap.begin()->second;
+					layerNames = settings.layerNames;
+				}
+
+				// Layer Bits (이 오브젝트가 속한 레이어) - 1개만 선택 가능
+				// 주의: PhysX 엔진은 32개 레이어를 지원하지만, 에디터 UI는 편의상 16개만 표시합니다.
+				ImGui::Text("Layer");
+				ImGui::Indent();
+				{
+					// 현재 선택된 레이어 찾기
+					int currentLayer = -1;
+					for (int i = 0; i < 16; ++i) // 16개만 확인 (UI 제한)
+					{
+						if ((collider->layerBits & (1u << i)) != 0)
+						{
+							currentLayer = i;
+							break;
+						}
+					}
+
+					// 16개 이상의 레이어가 선택되어 있으면 초기화
+					if (currentLayer == -1 && collider->layerBits != 0)
+					{
+						collider->layerBits = 0;
+						changed = true;
+					}
+
+					// ComboBox로 레이어 선택
+					std::string preview = (currentLayer >= 0) ?
+						(layerNames[currentLayer].empty() ? ("Layer " + std::to_string(currentLayer)) : layerNames[currentLayer]) :
+						"None";
+
+					if (ImGui::BeginCombo("##LayerBits", preview.c_str()))
+					{
+						if (ImGui::Selectable("None", currentLayer == -1))
+						{
+							collider->layerBits = 0;
+							changed = true;
+						}
+						for (int i = 0; i < 16; ++i) // 16개만 표시
+						{
+							std::string layerName = layerNames[i].empty() ? ("Layer " + std::to_string(i)) : layerNames[i];
+							bool isSelected = (currentLayer == i);
+							if (ImGui::Selectable(layerName.c_str(), isSelected))
+							{
+								collider->layerBits = (1u << i); // 단일 레이어만 설정
+								changed = true;
+							}
+							if (isSelected)
+								ImGui::SetItemDefaultFocus();
+						}
+						ImGui::EndCombo();
+					}
+				}
+				ImGui::Unindent();
+
+				// collideMask/queryMask는 레이어 매트릭스로만 결정됨 (컴포넌트에서 제거됨)
+				ImGui::TextDisabled("(Collide/Query Mask는 Physics Scene Settings의 레이어 매트릭스로 결정됩니다)");
+
+				// Ignore Layers (칩 UI)
+				ImGui::Text("Ignore Layers");
+				ImGui::Indent();
+				changed |= DrawIgnoreLayersChipEditor("IgnoreLayers", collider->ignoreLayers, layerNames);
+				ImGui::Unindent();
+
+				if (changed) g_SceneDirty = true;
+			}
+		}
+	}
+
+
+void EditorCore::DrawInspectorMeshCollider(World& world, const EntityId& _selectedEntity)
+	{
+		if (auto* meshCollider = world.GetComponent<Phy_MeshColliderComponent>(_selectedEntity))
+		{
+			if (ImGui::CollapsingHeader("Mesh Collider", ImGuiTreeNodeFlags_DefaultOpen))
+			{
+				bool changed = false;
+
+				if (ImGui::Button("Remove##MeshColliderRemove"))
+				{
+					world.RemoveComponent<Phy_MeshColliderComponent>(_selectedEntity);
+					g_SceneDirty = true;
+					return;
+				}
+
+				changed |= ImGui::Checkbox("Debug Draw", &meshCollider->debugDraw);
+
+				// Mesh Collider Type 선택
+				ImGui::Text("Mesh Collider Type");
+				ImGui::Indent();
+				{
+					const char* typeLabels[] = { "Triangle", "Convex" };
+					int typeIndex = static_cast<int>(meshCollider->type);
+					if (ImGui::Combo("##MeshColliderType", &typeIndex, typeLabels, IM_ARRAYSIZE(typeLabels)))
+					{
+						meshCollider->type = static_cast<MeshColliderType>(typeIndex);
+						changed = true;
+					}
+				}
+				ImGui::Unindent();
+
+				// MeshCollider 편집 이벤트 처리
+				static EntityId lastEditedMeshColliderEntity = InvalidEntityId;
+				static JsonRttr::json meshColliderEditStartJson;
+
+				ReflectionUI::UIEditEvent event = ReflectionUI::RenderInspector(*meshCollider, [](const std::string& name) {
+					return name != "type" && name != "layerBits" && name != "collideMask" && name != "queryMask" &&
+						name != "ignoreLayers" && name != "physicsActorHandle" &&
+						name != "flipNormals" && name != "doubleSidedQueries" && name != "validate" &&
+						name != "shiftVertices" && name != "vertexLimit";
+				});
+
+				// 편집 시작
+				if (event.activated && _selectedEntity != lastEditedMeshColliderEntity)
+				{
+					rttr::instance inst = *meshCollider;
+					meshColliderEditStartJson = JsonRttr::ToJsonObject(inst);
+					lastEditedMeshColliderEntity = _selectedEntity;
+				}
+
+				// 편집 종료
+				if (event.deactivatedAfterEdit && _selectedEntity == lastEditedMeshColliderEntity)
+				{
+					rttr::instance inst = *meshCollider;
+					JsonRttr::json meshColliderEditEndJson = JsonRttr::ToJsonObject(inst);
+
+					if (meshColliderEditStartJson != meshColliderEditEndJson)
+					{
+						Phy_MeshColliderComponent oldMeshCollider, newMeshCollider;
+						rttr::instance oldInst = oldMeshCollider;
+						rttr::instance newInst = newMeshCollider;
+						JsonRttr::FromJsonObject(oldInst, meshColliderEditStartJson);
+						JsonRttr::FromJsonObject(newInst, meshColliderEditEndJson);
+						PushCommand(std::make_unique<ComponentEditCommand<Phy_MeshColliderComponent>>(_selectedEntity, oldMeshCollider, newMeshCollider));
+						g_SceneDirty = true;
+					}
+
+					lastEditedMeshColliderEntity = InvalidEntityId;
+				}
+
+				changed |= event.changed;
+
+				ImGui::Separator();
+				ImGui::TextUnformatted("Mesh Options");
+
+				// Mesh Asset 선택 (ComboBox로 이미 로드된 메시 목록 표시)
+				ImGui::Text("Mesh Asset");
+				ImGui::Indent();
+				{
+					// 현재 선택된 메시 경로
+					std::string currentPath = meshCollider->meshAssetPath;
+
+					// 동일 엔티티의 SkinnedMeshComponent 확인
+					const auto* skinned = world.GetComponent<SkinnedMeshComponent>(_selectedEntity);
+					bool useSkinnedMesh = currentPath.empty() && skinned && !skinned->meshAssetPath.empty();
+
+					if (useSkinnedMesh)
+						currentPath = skinned->meshAssetPath;
+
+					// Preview 텍스트
+					std::string preview = "Auto (Use SkinnedMeshComponent)";
+					if (!currentPath.empty())
+						preview = currentPath;
+
+					if (ImGui::BeginCombo("##MeshAssetPath", preview.c_str()))
+					{
+						// Auto 옵션 (SkinnedMeshComponent 사용)
+						bool isAuto = meshCollider->meshAssetPath.empty();
+						if (ImGui::Selectable("Auto (Use SkinnedMeshComponent)", isAuto))
+						{
+							meshCollider->meshAssetPath.clear();
+							changed = true;
+						}
+						if (isAuto)
+							ImGui::SetItemDefaultFocus();
+
+						// 이미 로드된 메시 목록 표시 (SkinnedMeshRegistry에서)
+						if (m_skinnedRegistry)
+						{
+							// SkinnedMeshComponent가 있는 모든 엔티티를 순회하여 등록된 메시 수집
+							std::set<std::string> registeredMeshes;
+							auto skinnedComps = world.GetComponents<SkinnedMeshComponent>();
+							for (const auto& [eid, comp] : skinnedComps)
+							{
+								if (!comp.meshAssetPath.empty())
+								{
+									// 레지스트리에 실제로 등록되어 있는지 확인
+									if (m_skinnedRegistry->Find(comp.meshAssetPath))
+										registeredMeshes.insert(comp.meshAssetPath);
+								}
+							}
+
+							// 등록된 메시 목록 표시
+							for (const auto& meshPath : registeredMeshes)
+							{
+								bool isSelected = (currentPath == meshPath);
+								if (ImGui::Selectable(meshPath.c_str(), isSelected))
+								{
+									meshCollider->meshAssetPath = meshPath;
+									changed = true;
+								}
+								if (isSelected)
+									ImGui::SetItemDefaultFocus();
+							}
+						}
+						else
+						{
+							ImGui::TextDisabled("(SkinnedMeshRegistry not available)");
+						}
+
+						ImGui::EndCombo();
+					}
+
+					// 현재 상태 표시
+					if (meshCollider->meshAssetPath.empty())
+					{
+						if (skinned && !skinned->meshAssetPath.empty())
+						{
+							ImGui::TextColored(ImVec4(0.0f, 1.0f, 0.0f, 1.0f),
+								"Using: %s", skinned->meshAssetPath.c_str());
+						}
+						else
+						{
+							ImGui::TextColored(ImVec4(1.0f, 0.5f, 0.0f, 1.0f),
+								"No SkinnedMeshComponent found on this entity");
+						}
+					}
+				}
+				ImGui::Unindent();
+
+				if (meshCollider->type == MeshColliderType::Triangle)
+				{
+					changed |= ImGui::Checkbox("Flip Normals", &meshCollider->flipNormals);
+					changed |= ImGui::Checkbox("Double-Sided Queries", &meshCollider->doubleSidedQueries);
+					changed |= ImGui::Checkbox("Validate (Debug)", &meshCollider->validate);
+				}
+				else
+				{
+					changed |= ImGui::Checkbox("Shift Vertices", &meshCollider->shiftVertices);
+					changed |= ImGui::InputScalar("Vertex Limit", ImGuiDataType_U32, &meshCollider->vertexLimit);
+					changed |= ImGui::Checkbox("Validate (Debug)", &meshCollider->validate);
+				}
+
+				// 레이어 마스크 편집
+				ImGui::Separator();
+				ImGui::Text("Layer Settings");
+
+				std::array<std::string, 32> layerNames;
+				for (int i = 0; i < 32; ++i)
+					layerNames[i] = "Layer " + std::to_string(i);
+
+				const auto& settingsMap = world.GetComponents<Phy_SettingsComponent>();
+				if (!settingsMap.empty())
+				{
+					const auto& settings = settingsMap.begin()->second;
+					layerNames = settings.layerNames;
+				}
+
+				ImGui::Text("Layer");
+				ImGui::Indent();
+				{
+					int currentLayer = -1;
+					for (int i = 0; i < 16; ++i)
+					{
+						if ((meshCollider->layerBits & (1u << i)) != 0)
+						{
+							currentLayer = i;
+							break;
+						}
+					}
+
+					if (currentLayer == -1 && meshCollider->layerBits != 0)
+					{
+						meshCollider->layerBits = 0;
+						changed = true;
+					}
+
+					std::string preview = (currentLayer >= 0) ?
+						(layerNames[currentLayer].empty() ? ("Layer " + std::to_string(currentLayer)) : layerNames[currentLayer]) :
+						"None";
+
+					if (ImGui::BeginCombo("##MeshColliderLayerBits", preview.c_str()))
+					{
+						if (ImGui::Selectable("None", currentLayer == -1))
+						{
+							meshCollider->layerBits = 0;
+							changed = true;
+						}
+						for (int i = 0; i < 16; ++i)
+						{
+							std::string layerName = layerNames[i].empty() ? ("Layer " + std::to_string(i)) : layerNames[i];
+							bool isSelected = (currentLayer == i);
+							if (ImGui::Selectable(layerName.c_str(), isSelected))
+							{
+								meshCollider->layerBits = (1u << i);
+								changed = true;
+							}
+							if (isSelected)
+								ImGui::SetItemDefaultFocus();
+						}
+						ImGui::EndCombo();
+					}
+				}
+				ImGui::Unindent();
+
+				// collideMask/queryMask는 레이어 매트릭스로만 결정됨 (컴포넌트에서 제거됨)
+				ImGui::TextDisabled("(Collide/Query Mask는 Physics Scene Settings의 레이어 매트릭스로 결정됩니다)");
+
+				ImGui::Text("Ignore Layers");
+				ImGui::Indent();
+				changed |= DrawIgnoreLayersChipEditor("MeshIgnoreLayers", meshCollider->ignoreLayers, layerNames);
+				ImGui::Unindent();
+
+				if (changed) g_SceneDirty = true;
+			}
+		}
+	}
+
+
+void EditorCore::DrawInspectorCharacterController(World& world, const EntityId& _selectedEntity)
+	{
+		if (auto* cct = world.GetComponent<Phy_CCTComponent>(_selectedEntity))
+		{
+			if (ImGui::CollapsingHeader("Character Controller", ImGuiTreeNodeFlags_DefaultOpen))
+			{
+				bool changed = false;
+
+				if (ImGui::Button("Remove##CCTRemove"))
+				{
+					world.RemoveComponent<Phy_CCTComponent>(_selectedEntity);
+					g_SceneDirty = true;
+					return;
+				}
+
+				// 기본 프로퍼티는 ReflectionUI로
+				// CCT 편집 이벤트 처리
+				static EntityId lastEditedCCTEntity = InvalidEntityId;
+				static JsonRttr::json cctEditStartJson;
+
+				ReflectionUI::UIEditEvent event = ReflectionUI::RenderInspector(*cct, [](const std::string& name) {
+					// layerBits는 커스텀 UI로 처리 (collideMask/queryMask는 레이어 매트릭스로만 결정)
+					return name != "layerBits" && name != "controllerHandle";
+				});
+
+				// 편집 시작
+				if (event.activated && _selectedEntity != lastEditedCCTEntity)
+				{
+					rttr::instance inst = *cct;
+					cctEditStartJson = JsonRttr::ToJsonObject(inst);
+					lastEditedCCTEntity = _selectedEntity;
+				}
+
+				// 편집 종료
+				if (event.deactivatedAfterEdit && _selectedEntity == lastEditedCCTEntity)
+				{
+					rttr::instance inst = *cct;
+					JsonRttr::json cctEditEndJson = JsonRttr::ToJsonObject(inst);
+
+					if (cctEditStartJson != cctEditEndJson)
+					{
+						Phy_CCTComponent oldCCT, newCCT;
+						rttr::instance oldInst = oldCCT;
+						rttr::instance newInst = newCCT;
+						JsonRttr::FromJsonObject(oldInst, cctEditStartJson);
+						JsonRttr::FromJsonObject(newInst, cctEditEndJson);
+						PushCommand(std::make_unique<ComponentEditCommand<Phy_CCTComponent>>(_selectedEntity, oldCCT, newCCT));
+						g_SceneDirty = true;
+					}
+
+					lastEditedCCTEntity = InvalidEntityId;
+				}
+
+				changed |= event.changed;
+
+				// 레이어 마스크 편집
+				ImGui::Separator();
+				ImGui::Text("Layer Settings");
+
+				// Phy_SettingsComponent에서 레이어 이름 가져오기
+				std::array<std::string, 32> layerNames;
+				for (int i = 0; i < 32; ++i)
+					layerNames[i] = "Layer " + std::to_string(i);
+
+				const auto& settingsMap = world.GetComponents<Phy_SettingsComponent>();
+				if (!settingsMap.empty())
+				{
+					const auto& settings = settingsMap.begin()->second;
+					layerNames = settings.layerNames;
+				}
+
+				// Layer Bits (이 오브젝트가 속한 레이어) - 1개만 선택 가능
+				// 주의: PhysX 엔진은 32개 레이어를 지원하지만, 에디터 UI는 편의상 16개만 표시합니다.
+				ImGui::Text("Layer");
+				ImGui::Indent();
+				{
+					// 현재 선택된 레이어 찾기
+					int currentLayer = -1;
+					for (int i = 0; i < 16; ++i) // 16개만 확인 (UI 제한)
+					{
+						if ((cct->layerBits & (1u << i)) != 0)
+						{
+							currentLayer = i;
+							break;
+						}
+					}
+
+					// 16개 이상의 레이어가 선택되어 있으면 초기화
+					if (currentLayer == -1 && cct->layerBits != 0)
+					{
+						cct->layerBits = 0;
+						changed = true;
+					}
+
+					// ComboBox로 레이어 선택
+					std::string preview = (currentLayer >= 0) ?
+						(layerNames[currentLayer].empty() ? ("Layer " + std::to_string(currentLayer)) : layerNames[currentLayer]) :
+						"None";
+
+					if (ImGui::BeginCombo("##LayerBits", preview.c_str()))
+					{
+						if (ImGui::Selectable("None", currentLayer == -1))
+						{
+							cct->layerBits = 0;
+							changed = true;
+						}
+						for (int i = 0; i < 16; ++i) // 16개만 표시
+						{
+							std::string layerName = layerNames[i].empty() ? ("Layer " + std::to_string(i)) : layerNames[i];
+							bool isSelected = (currentLayer == i);
+							if (ImGui::Selectable(layerName.c_str(), isSelected))
+							{
+								cct->layerBits = (1u << i); // 단일 레이어만 설정
+								changed = true;
+							}
+							if (isSelected)
+								ImGui::SetItemDefaultFocus();
+						}
+						ImGui::EndCombo();
+					}
+				}
+				ImGui::Unindent();
+
+				// collideMask/queryMask는 레이어 매트릭스로만 결정됨 (컴포넌트에서 제거됨)
+				ImGui::TextDisabled("(Collide/Query Mask는 Physics Scene Settings의 레이어 매트릭스로 결정됩니다)");
+
+				// Ignore Layers (칩 UI)
+				ImGui::Text("Ignore Layers");
+				ImGui::Indent();
+				changed |= DrawIgnoreLayersChipEditor("IgnoreLayers", cct->ignoreLayers, layerNames);
+				ImGui::Unindent();
+
+				if (changed) g_SceneDirty = true;
+			}
+		}
+	}
+
+
+void EditorCore::DrawInspectorPhysicsSceneSettings(World& world, const EntityId& _selectedEntity)
+	{
+		if (auto* settings = world.GetComponent<Phy_SettingsComponent>(_selectedEntity))
+		{
+			if (ImGui::CollapsingHeader("Physics Scene Settings", ImGuiTreeNodeFlags_DefaultOpen))
+			{
+				bool changed = false;
+
+				if (ImGui::Button("Remove##PhysicsSettingsRemove"))
+				{
+					world.RemoveComponent<Phy_SettingsComponent>(_selectedEntity);
+					g_SceneDirty = true;
+					return;
+				}
+
+				if (ImGui::Button("Apply Combat Defaults"))
+				{
+					CombatPhysicsLayers::ApplyDefaultCombatLayerMatrix(*settings);
+					settings->filterRevision++;
+					changed = true;
+					g_SceneDirty = true;
+				}
+
+				// 기본 프로퍼티는 ReflectionUI로
+				// PhysicsSettings 편집 이벤트 처리
+				static EntityId lastEditedPhysicsSettingsEntity = InvalidEntityId;
+				static JsonRttr::json physicsSettingsEditStartJson;
+
+				ReflectionUI::UIEditEvent event = ReflectionUI::RenderInspector(*settings, [](const std::string& name) {
+					// layerCollideMatrix, layerQueryMatrix, layerNames는 커스텀 UI로 처리
+					return name != "layerCollideMatrix" && name != "layerQueryMatrix" && name != "layerNames" &&
+						name != "enableGroundPlane" && name != "groundStaticFriction" && name != "groundDynamicFriction" &&
+						name != "groundRestitution" && name != "groundLayerBits" && name != "groundCollideMask" &&
+						name != "groundQueryMask" && name != "groundIgnoreLayers" && name != "groundIsTrigger";
+				});
+
+				// 편집 시작
+				if (event.activated && _selectedEntity != lastEditedPhysicsSettingsEntity)
+				{
+					rttr::instance inst = *settings;
+					physicsSettingsEditStartJson = JsonRttr::ToJsonObject(inst);
+					lastEditedPhysicsSettingsEntity = _selectedEntity;
+				}
+
+				// 편집 종료
+				if (event.deactivatedAfterEdit && _selectedEntity == lastEditedPhysicsSettingsEntity)
+				{
+					rttr::instance inst = *settings;
+					JsonRttr::json physicsSettingsEditEndJson = JsonRttr::ToJsonObject(inst);
+
+					if (physicsSettingsEditStartJson != physicsSettingsEditEndJson)
+					{
+						Phy_SettingsComponent oldSettings, newSettings;
+						rttr::instance oldInst = oldSettings;
+						rttr::instance newInst = newSettings;
+						JsonRttr::FromJsonObject(oldInst, physicsSettingsEditStartJson);
+						JsonRttr::FromJsonObject(newInst, physicsSettingsEditEndJson);
+						PushCommand(std::make_unique<ComponentEditCommand<Phy_SettingsComponent>>(_selectedEntity, oldSettings, newSettings));
+						g_SceneDirty = true;
+					}
+
+					lastEditedPhysicsSettingsEntity = InvalidEntityId;
+				}
+
+				changed |= event.changed;
+
+				ImGui::Separator();
+				ImGui::Text("Ground Plane (y=0)");
+				changed |= ImGui::Checkbox("Enable Ground Plane", &settings->enableGroundPlane);
+
+				if (settings->enableGroundPlane)
+				{
+					changed |= ImGui::DragFloat("Static Friction", &settings->groundStaticFriction, 0.01f, 0.0f, 10.0f);
+					changed |= ImGui::DragFloat("Dynamic Friction", &settings->groundDynamicFriction, 0.01f, 0.0f, 10.0f);
+					changed |= ImGui::DragFloat("Restitution", &settings->groundRestitution, 0.01f, 0.0f, 1.0f);
+					changed |= ImGui::Checkbox("Trigger", &settings->groundIsTrigger);
+
+					std::array<std::string, 32> layerNames = settings->layerNames;
+
+					ImGui::Text("Layer");
+					ImGui::Indent();
+					{
+						int currentLayer = -1;
+						for (int i = 0; i < 16; ++i)
+						{
+							if ((settings->groundLayerBits & (1u << i)) != 0)
+							{
+								currentLayer = i;
+								break;
+							}
+						}
+
+						if (currentLayer == -1 && settings->groundLayerBits != 0)
+						{
+							settings->groundLayerBits = 0;
+							changed = true;
+						}
+
+						std::string preview = (currentLayer >= 0) ?
+							(layerNames[currentLayer].empty() ? ("Layer " + std::to_string(currentLayer)) : layerNames[currentLayer]) :
+							"None";
+
+						if (ImGui::BeginCombo("##GroundLayerBits", preview.c_str()))
+						{
+							if (ImGui::Selectable("None", currentLayer == -1))
+							{
+								settings->groundLayerBits = 0;
+								changed = true;
+							}
+							for (int i = 0; i < 16; ++i)
+							{
+								std::string layerName = layerNames[i].empty() ? ("Layer " + std::to_string(i)) : layerNames[i];
+								bool isSelected = (currentLayer == i);
+								if (ImGui::Selectable(layerName.c_str(), isSelected))
+								{
+									settings->groundLayerBits = (1u << i);
+									changed = true;
+								}
+								if (isSelected)
+									ImGui::SetItemDefaultFocus();
+							}
+							ImGui::EndCombo();
+						}
+					}
+					ImGui::Unindent();
+
+					changed |= DrawLayerMaskChipEditor("Ground Collide Mask", settings->groundCollideMask, layerNames);
+					changed |= DrawLayerMaskChipEditor("Ground Query Mask", settings->groundQueryMask, layerNames);
+
+					ImGui::Text("Ground Ignore Layers");
+					ImGui::Indent();
+					changed |= DrawIgnoreLayersChipEditor("GroundIgnoreLayers", settings->groundIgnoreLayers, layerNames);
+					ImGui::Unindent();
+				}
+
+				ImGui::Separator();
+				ImGui::Text("Layer Collision Matrix");
+				ImGui::Text("(Collide Mask: Check = Collision enabled between layers)");
+
+				// 레이어 충돌 매트릭스 편집
+				// 주의: PhysX 엔진은 32개 레이어를 지원하지만, 에디터 UI는 편의상 16개만 표시합니다.
+				ImGui::BeginChild("LayerCollideMatrix", ImVec2(0, 400), false, ImGuiWindowFlags_HorizontalScrollbar);
+
+				// 행 단위로 표시: "00 | 00 [ ] 01 [ ] 02 [ ] 03 [ ]"
+				for (int i = 0; i < 16; ++i)
+				{
+					// 행 번호 표시 (2자리로 포맷팅)
+					char rowLabel[8];
+					snprintf(rowLabel, sizeof(rowLabel), "%02d |", i);
+					ImGui::Text("%s", rowLabel);
+					ImGui::SameLine();
+
+					// 해당 행의 모든 열에 대한 체크박스 표시
+					for (int j = 0; j < 16; ++j)
+					{
+						bool collision = settings->layerCollideMatrix[i][j];
+						char colLabel[8];
+						snprintf(colLabel, sizeof(colLabel), "%02d", j);
+						ImGui::PushID(i * 16 + j);
+						if (ImGui::Checkbox(colLabel, &collision))
+						{
+							settings->layerCollideMatrix[i][j] = collision;
+							settings->layerCollideMatrix[j][i] = collision; // 충돌은 대칭 (필수)
+							settings->filterRevision++; // 필터 변경 감지용
+							changed = true;
+						}
+						ImGui::PopID();
+						ImGui::SameLine();
+					}
+					ImGui::NewLine();
+				}
+
+				ImGui::EndChild();
+
+				ImGui::Separator();
+				ImGui::Text("Layer Query Matrix");
+				ImGui::Text("(Query Mask: Check = Query enabled between layers)");
+
+				// 레이어 쿼리 매트릭스 편집
+				// 주의: PhysX 엔진은 32개 레이어를 지원하지만, 에디터 UI는 편의상 16개만 표시합니다.
+				ImGui::BeginChild("LayerQueryMatrix", ImVec2(0, 400), false, ImGuiWindowFlags_HorizontalScrollbar);
+
+				// 행 단위로 표시: "00 | 00 [ ] 01 [ ] 02 [ ] 03 [ ]"
+				for (int i = 0; i < 16; ++i)
+				{
+					// 행 번호 표시 (2자리로 포맷팅)
+					char rowLabel[8];
+					snprintf(rowLabel, sizeof(rowLabel), "%02d |", i);
+					ImGui::Text("%s", rowLabel);
+					ImGui::SameLine();
+
+					// 해당 행의 모든 열에 대한 체크박스 표시
+					for (int j = 0; j < 16; ++j)
+					{
+						bool query = settings->layerQueryMatrix[i][j];
+						char colLabel[8];
+						snprintf(colLabel, sizeof(colLabel), "%02d", j);
+						ImGui::PushID(10000 + i * 16 + j);
+						if (ImGui::Checkbox(colLabel, &query))
+						{
+							settings->layerQueryMatrix[i][j] = query;
+							// 쿼리는 비대칭이 유용한 경우가 많으므로 대칭 적용 제거
+							// (예: 카메라 레이는 특정 레이어만 보고, AI는 또 다르게 봄)
+							settings->filterRevision++; // 필터 변경 감지용
+							changed = true;
+						}
+						ImGui::PopID();
+						ImGui::SameLine();
+					}
+					ImGui::NewLine();
+				}
+
+				ImGui::EndChild();
+
+				if (changed) g_SceneDirty = true;
+			}
+		}
+	}
+
+
+void EditorCore::DrawInspectorTerrainHeightField(World& world, const EntityId& _selectedEntity)
+	{
+		if (auto* terrain = world.GetComponent<Phy_TerrainHeightFieldComponent>(_selectedEntity))
+		{
+			if (ImGui::CollapsingHeader("Terrain Height Field", ImGuiTreeNodeFlags_DefaultOpen))
+			{
+				bool changed = false;
+
+				if (ImGui::Button("Remove##TerrainRemove"))
+				{
+					world.RemoveComponent<Phy_TerrainHeightFieldComponent>(_selectedEntity);
+					g_SceneDirty = true;
+					return;
+				}
+
+				// 기본 프로퍼티는 ReflectionUI로
+				// Terrain 편집 이벤트 처리
+				static EntityId lastEditedTerrainEntity = InvalidEntityId;
+				static JsonRttr::json terrainEditStartJson;
+
+				ReflectionUI::UIEditEvent event = ReflectionUI::RenderInspector(*terrain, [](const std::string& name) {
+					// layerBits, heightSamples는 커스텀 UI로 처리 (collideMask/queryMask는 레이어 매트릭스로만 결정)
+					return name != "layerBits" && name != "heightSamples" && name != "physicsActorHandle";
+				});
+
+				// 편집 시작
+				if (event.activated && _selectedEntity != lastEditedTerrainEntity)
+				{
+					rttr::instance inst = *terrain;
+					terrainEditStartJson = JsonRttr::ToJsonObject(inst);
+					lastEditedTerrainEntity = _selectedEntity;
+				}
+
+				// 편집 종료
+				if (event.deactivatedAfterEdit && _selectedEntity == lastEditedTerrainEntity)
+				{
+					rttr::instance inst = *terrain;
+					JsonRttr::json terrainEditEndJson = JsonRttr::ToJsonObject(inst);
+
+					if (terrainEditStartJson != terrainEditEndJson)
+					{
+						Phy_TerrainHeightFieldComponent oldTerrain, newTerrain;
+						rttr::instance oldInst = oldTerrain;
+						rttr::instance newInst = newTerrain;
+						JsonRttr::FromJsonObject(oldInst, terrainEditStartJson);
+						JsonRttr::FromJsonObject(newInst, terrainEditEndJson);
+						PushCommand(std::make_unique<ComponentEditCommand<Phy_TerrainHeightFieldComponent>>(_selectedEntity, oldTerrain, newTerrain));
+						g_SceneDirty = true;
+					}
+
+					lastEditedTerrainEntity = InvalidEntityId;
+				}
+
+				changed |= event.changed;
+
+				// HeightSamples 상태 표시 및 생성 버튼
+				ImGui::Separator();
+				ImGui::Text("Height Samples");
+				ImGui::Indent();
+				{
+					const size_t expectedSamples = static_cast<size_t>(terrain->numRows) * static_cast<size_t>(terrain->numCols);
+					const bool isValid = (terrain->numRows >= 2 && terrain->numCols >= 2) &&
+						(terrain->heightSamples.size() == expectedSamples);
+
+					if (terrain->heightSamples.empty())
+					{
+						ImGui::TextColored(ImVec4(1.0f, 0.5f, 0.0f, 1.0f), "Status: Empty (requires data)");
+					}
+					else if (!isValid)
+					{
+						ImGui::TextColored(ImVec4(1.0f, 0.0f, 0.0f, 1.0f),
+							"Status: Invalid (size: %zu, expected: %zu)",
+							terrain->heightSamples.size(), expectedSamples);
+					}
+					else
+					{
+						ImGui::TextColored(ImVec4(0.0f, 1.0f, 0.0f, 1.0f),
+							"Status: Valid (size: %zu)",
+							terrain->heightSamples.size());
+					}
+
+					ImGui::Text("Grid: %u x %u (total: %zu samples)",
+						terrain->numRows, terrain->numCols, expectedSamples);
+
+					if (terrain->numRows >= 2 && terrain->numCols >= 2)
+					{
+						if (ImGui::Button("Generate Flat (0.0)"))
+						{
+							terrain->heightSamples.resize(expectedSamples, 0.0f);
+							changed = true;
+							g_SceneDirty = true;
+						}
+						ImGui::SameLine();
+						ImGui::TextDisabled("(?)");
+						if (ImGui::IsItemHovered())
+						{
+							ImGui::BeginTooltip();
+							ImGui::Text("Generates a flat terrain with all heights set to 0.0");
+							ImGui::EndTooltip();
+						}
+					}
+					else
+					{
+						ImGui::TextDisabled("Set numRows and numCols (>= 2) to enable generation");
+					}
+				}
+				ImGui::Unindent();
+
+				// 레이어 마스크 편집
+				ImGui::Separator();
+				ImGui::Text("Layer Settings");
+
+				// Phy_SettingsComponent에서 레이어 이름 가져오기
+				std::array<std::string, 32> layerNames;
+				for (int i = 0; i < 32; ++i)
+					layerNames[i] = "Layer " + std::to_string(i);
+
+				const auto& settingsMap = world.GetComponents<Phy_SettingsComponent>();
+				if (!settingsMap.empty())
+				{
+					const auto& settings = settingsMap.begin()->second;
+					layerNames = settings.layerNames;
+				}
+
+				// Layer Bits (이 오브젝트가 속한 레이어) - 1개만 선택 가능
+				// 주의: PhysX 엔진은 32개 레이어를 지원하지만, 에디터 UI는 편의상 16개만 표시합니다.
+				ImGui::Text("Layer");
+				ImGui::Indent();
+				{
+					// 현재 선택된 레이어 찾기
+					int currentLayer = -1;
+					for (int i = 0; i < 16; ++i) // 16개만 확인 (UI 제한)
+					{
+						if ((terrain->layerBits & (1u << i)) != 0)
+						{
+							currentLayer = i;
+							break;
+						}
+					}
+
+					// 16개 이상의 레이어가 선택되어 있으면 초기화
+					if (currentLayer == -1 && terrain->layerBits != 0)
+					{
+						terrain->layerBits = 0;
+						changed = true;
+					}
+
+					// ComboBox로 레이어 선택
+					std::string preview = (currentLayer >= 0) ?
+						(layerNames[currentLayer].empty() ? ("Layer " + std::to_string(currentLayer)) : layerNames[currentLayer]) :
+						"None";
+
+					if (ImGui::BeginCombo("##LayerBits", preview.c_str()))
+					{
+						if (ImGui::Selectable("None", currentLayer == -1))
+						{
+							terrain->layerBits = 0;
+							changed = true;
+						}
+						for (int i = 0; i < 16; ++i) // 16개만 표시
+						{
+							std::string layerName = layerNames[i].empty() ? ("Layer " + std::to_string(i)) : layerNames[i];
+							bool isSelected = (currentLayer == i);
+							if (ImGui::Selectable(layerName.c_str(), isSelected))
+							{
+								terrain->layerBits = (1u << i); // 단일 레이어만 설정
+								changed = true;
+							}
+							if (isSelected)
+								ImGui::SetItemDefaultFocus();
+						}
+						ImGui::EndCombo();
+					}
+				}
+				ImGui::Unindent();
+
+				// collideMask/queryMask는 레이어 매트릭스로만 결정됨 (컴포넌트에서 제거됨)
+				ImGui::TextDisabled("(Collide/Query Mask는 Physics Scene Settings의 레이어 매트릭스로 결정됩니다)");
+
+				// Ignore Layers (칩 UI)
+				ImGui::Text("Ignore Layers");
+				ImGui::Indent();
+				changed |= DrawIgnoreLayersChipEditor("IgnoreLayers", terrain->ignoreLayers, layerNames);
+				ImGui::Unindent();
+
+				if (changed) g_SceneDirty = true;
+			}
+		}
+	}
+
+
+void EditorCore::DrawInspectorJoint(World& world, const EntityId& _selectedEntity)
+	{
+		if (auto* joint = world.GetComponent<Phy_JointComponent>(_selectedEntity))
+		{
+			if (ImGui::CollapsingHeader("Joint", ImGuiTreeNodeFlags_DefaultOpen))
+			{
+				bool changed = false;
+
+				if (ImGui::Button("Remove##JointRemove"))
+				{
+					world.RemoveComponent<Phy_JointComponent>(_selectedEntity);
+					g_SceneDirty = true;
+					return;
+				}
+
+				const char* typeLabels[] = { "Fixed", "Revolute", "Prismatic", "Distance", "Spherical", "D6" };
+				int typeIndex = static_cast<int>(joint->type);
+				if (ImGui::Combo("Type", &typeIndex, typeLabels, IM_ARRAYSIZE(typeLabels)))
+				{
+					joint->type = static_cast<Phy_JointType>(typeIndex);
+					changed = true;
+				}
+
+				// Target Entity 선택 (ComboBox)
+				ImGui::Text("Target Entity");
+				ImGui::Indent();
+				{
+					// 현재 타겟 엔티티 찾기
+					EntityId currentTargetId = Alice::InvalidEntityId;
+					std::string currentTargetName = joint->targetName;
+					if (!currentTargetName.empty())
+					{
+						GameObject targetGo = world.FindGameObject(currentTargetName);
+						if (targetGo.IsValid())
+							currentTargetId = targetGo.id();
+					}
+
+					// Preview 텍스트 생성
+					std::string preview = "None";
+					if (currentTargetId != Alice::InvalidEntityId)
+					{
+						std::string name = world.GetEntityName(currentTargetId);
+						if (name.empty())
+							name = "Entity " + std::to_string((uint32_t)currentTargetId);
+						preview = name;
+					}
+					else if (!currentTargetName.empty())
+					{
+						preview = currentTargetName + " (not found)";
+					}
+
+					if (ImGui::BeginCombo("##JointTarget", preview.c_str()))
+					{
+						// None 옵션
+						if (ImGui::Selectable("None", currentTargetId == Alice::InvalidEntityId))
+						{
+							joint->targetName.clear();
+							changed = true;
+						}
+						if (currentTargetId == Alice::InvalidEntityId)
+							ImGui::SetItemDefaultFocus();
+
+						// 모든 엔티티 나열
+						auto transforms = world.GetComponents<TransformComponent>();
+						for (const auto& [entityId, transform] : transforms)
+						{
+							std::string name = world.GetEntityName(entityId);
+							if (name.empty())
+								name = "Entity " + std::to_string((uint32_t)entityId);
+
+							bool isSelected = (currentTargetId == entityId);
+							if (ImGui::Selectable(name.c_str(), isSelected))
+							{
+								joint->targetName = name;
+								changed = true;
+							}
+							if (isSelected)
+								ImGui::SetItemDefaultFocus();
+						}
+
+						ImGui::EndCombo();
+					}
+				}
+				ImGui::Unindent();
+
+				ImGui::Separator();
+				ImGui::Text("Common");
+				changed |= ImGui::Checkbox("Collide Connected", &joint->collideConnected);
+				changed |= ImGui::DragFloat("Break Force", &joint->breakForce, 1.0f, 0.0f);
+				changed |= ImGui::DragFloat("Break Torque", &joint->breakTorque, 1.0f, 0.0f);
+
+				auto drawFrame = [&](const char* label, Phy_JointFrame& frame) -> bool
+				{
+					bool frameChanged = false;
+					if (ImGui::TreeNode(label))
+					{
+						frameChanged |= ImGui::DragFloat3("Position", &frame.position.x, 0.01f);
+						frameChanged |= ImGui::DragFloat3("Rotation (Rad)", &frame.rotation.x, 0.01f);
+						ImGui::TreePop();
+					}
+					return frameChanged;
+				};
+
+				changed |= drawFrame("Frame A", joint->frameA);
+				changed |= drawFrame("Frame B", joint->frameB);
+
+				ImGui::Separator();
+				switch (joint->type)
+				{
+				case Phy_JointType::Fixed:
+					ImGui::Text("Fixed Joint: no extra settings");
+					break;
+				case Phy_JointType::Revolute:
+				{
+					if (ImGui::TreeNode("Revolute Limit"))
+					{
+						changed |= ImGui::Checkbox("Enable Limit", &joint->revolute.enableLimit);
+						changed |= ImGui::DragFloat("Lower Limit", &joint->revolute.lowerLimit, 0.01f);
+						changed |= ImGui::DragFloat("Upper Limit", &joint->revolute.upperLimit, 0.01f);
+						changed |= ImGui::DragFloat("Stiffness", &joint->revolute.limitStiffness, 0.01f);
+						changed |= ImGui::DragFloat("Damping", &joint->revolute.limitDamping, 0.01f);
+						changed |= ImGui::DragFloat("Restitution", &joint->revolute.limitRestitution, 0.01f);
+						changed |= ImGui::DragFloat("Bounce Threshold", &joint->revolute.limitBounceThreshold, 0.01f);
+						ImGui::TreePop();
+					}
+					if (ImGui::TreeNode("Revolute Drive"))
+					{
+						changed |= ImGui::Checkbox("Enable Drive", &joint->revolute.enableDrive);
+						changed |= ImGui::DragFloat("Drive Velocity", &joint->revolute.driveVelocity, 0.01f);
+						changed |= ImGui::DragFloat("Force Limit", &joint->revolute.driveForceLimit, 1.0f, 0.0f);
+						changed |= ImGui::Checkbox("Free Spin", &joint->revolute.driveFreeSpin);
+						changed |= ImGui::Checkbox("Drive Limits Are Forces", &joint->revolute.driveLimitsAreForces);
+						ImGui::TreePop();
+					}
+					break;
+				}
+				case Phy_JointType::Prismatic:
+				{
+					if (ImGui::TreeNode("Prismatic Limit"))
+					{
+						changed |= ImGui::Checkbox("Enable Limit", &joint->prismatic.enableLimit);
+						changed |= ImGui::DragFloat("Lower Limit", &joint->prismatic.lowerLimit, 0.01f);
+						changed |= ImGui::DragFloat("Upper Limit", &joint->prismatic.upperLimit, 0.01f);
+						changed |= ImGui::DragFloat("Stiffness", &joint->prismatic.limitStiffness, 0.01f);
+						changed |= ImGui::DragFloat("Damping", &joint->prismatic.limitDamping, 0.01f);
+						changed |= ImGui::DragFloat("Restitution", &joint->prismatic.limitRestitution, 0.01f);
+						changed |= ImGui::DragFloat("Bounce Threshold", &joint->prismatic.limitBounceThreshold, 0.01f);
+						ImGui::TreePop();
+					}
+					break;
+				}
+				case Phy_JointType::Distance:
+				{
+					if (ImGui::TreeNode("Distance"))
+					{
+						changed |= ImGui::DragFloat("Min Distance", &joint->distance.minDistance, 0.01f);
+						changed |= ImGui::DragFloat("Max Distance", &joint->distance.maxDistance, 0.01f);
+						changed |= ImGui::DragFloat("Tolerance", &joint->distance.tolerance, 0.01f);
+						changed |= ImGui::Checkbox("Enable Min", &joint->distance.enableMinDistance);
+						changed |= ImGui::Checkbox("Enable Max", &joint->distance.enableMaxDistance);
+						changed |= ImGui::Checkbox("Enable Spring", &joint->distance.enableSpring);
+						changed |= ImGui::DragFloat("Stiffness", &joint->distance.stiffness, 0.01f);
+						changed |= ImGui::DragFloat("Damping", &joint->distance.damping, 0.01f);
+						ImGui::TreePop();
+					}
+					break;
+				}
+				case Phy_JointType::Spherical:
+				{
+					if (ImGui::TreeNode("Spherical Limit"))
+					{
+						changed |= ImGui::Checkbox("Enable Limit", &joint->spherical.enableLimit);
+						changed |= ImGui::DragFloat("Y Limit Angle", &joint->spherical.yLimitAngle, 0.01f);
+						changed |= ImGui::DragFloat("Z Limit Angle", &joint->spherical.zLimitAngle, 0.01f);
+						changed |= ImGui::DragFloat("Stiffness", &joint->spherical.limitStiffness, 0.01f);
+						changed |= ImGui::DragFloat("Damping", &joint->spherical.limitDamping, 0.01f);
+						changed |= ImGui::DragFloat("Restitution", &joint->spherical.limitRestitution, 0.01f);
+						changed |= ImGui::DragFloat("Bounce Threshold", &joint->spherical.limitBounceThreshold, 0.01f);
+						ImGui::TreePop();
+					}
+					break;
+				}
+				case Phy_JointType::D6:
+				{
+					const char* motionLabels[] = { "Locked", "Limited", "Free" };
+					auto drawMotion = [&](const char* label, Phy_D6Motion& m)
+					{
+						int idx = static_cast<int>(m);
+						if (ImGui::Combo(label, &idx, motionLabels, IM_ARRAYSIZE(motionLabels)))
+						{
+							m = static_cast<Phy_D6Motion>(idx);
+							changed = true;
+						}
+					};
+
+					if (ImGui::TreeNode("Motions"))
+					{
+						drawMotion("Motion X", joint->d6.motionX);
+						drawMotion("Motion Y", joint->d6.motionY);
+						drawMotion("Motion Z", joint->d6.motionZ);
+						drawMotion("Motion Twist", joint->d6.motionTwist);
+						drawMotion("Motion Swing1", joint->d6.motionSwing1);
+						drawMotion("Motion Swing2", joint->d6.motionSwing2);
+						ImGui::TreePop();
+					}
+
+					if (ImGui::TreeNode("Linear Limits"))
+					{
+						ImGui::Text("X");
+						changed |= ImGui::DragFloat("Lower X", &joint->d6.linearLimitX.lower, 0.01f);
+						changed |= ImGui::DragFloat("Upper X", &joint->d6.linearLimitX.upper, 0.01f);
+						changed |= ImGui::DragFloat("Stiffness X", &joint->d6.linearLimitX.stiffness, 0.01f);
+						changed |= ImGui::DragFloat("Damping X", &joint->d6.linearLimitX.damping, 0.01f);
+						changed |= ImGui::DragFloat("Restitution X", &joint->d6.linearLimitX.restitution, 0.01f);
+						changed |= ImGui::DragFloat("Bounce Threshold X", &joint->d6.linearLimitX.bounceThreshold, 0.01f);
+						ImGui::Separator();
+
+						ImGui::Text("Y");
+						changed |= ImGui::DragFloat("Lower Y", &joint->d6.linearLimitY.lower, 0.01f);
+						changed |= ImGui::DragFloat("Upper Y", &joint->d6.linearLimitY.upper, 0.01f);
+						changed |= ImGui::DragFloat("Stiffness Y", &joint->d6.linearLimitY.stiffness, 0.01f);
+						changed |= ImGui::DragFloat("Damping Y", &joint->d6.linearLimitY.damping, 0.01f);
+						changed |= ImGui::DragFloat("Restitution Y", &joint->d6.linearLimitY.restitution, 0.01f);
+						changed |= ImGui::DragFloat("Bounce Threshold Y", &joint->d6.linearLimitY.bounceThreshold, 0.01f);
+						ImGui::Separator();
+
+						ImGui::Text("Z");
+						changed |= ImGui::DragFloat("Lower Z", &joint->d6.linearLimitZ.lower, 0.01f);
+						changed |= ImGui::DragFloat("Upper Z", &joint->d6.linearLimitZ.upper, 0.01f);
+						changed |= ImGui::DragFloat("Stiffness Z", &joint->d6.linearLimitZ.stiffness, 0.01f);
+						changed |= ImGui::DragFloat("Damping Z", &joint->d6.linearLimitZ.damping, 0.01f);
+						changed |= ImGui::DragFloat("Restitution Z", &joint->d6.linearLimitZ.restitution, 0.01f);
+						changed |= ImGui::DragFloat("Bounce Threshold Z", &joint->d6.linearLimitZ.bounceThreshold, 0.01f);
+						ImGui::TreePop();
+					}
+
+					if (ImGui::TreeNode("Angular Limits"))
+					{
+						ImGui::Text("Twist");
+						changed |= ImGui::DragFloat("Lower Twist", &joint->d6.twistLimit.lower, 0.01f);
+						changed |= ImGui::DragFloat("Upper Twist", &joint->d6.twistLimit.upper, 0.01f);
+						changed |= ImGui::DragFloat("Stiffness Twist", &joint->d6.twistLimit.stiffness, 0.01f);
+						changed |= ImGui::DragFloat("Damping Twist", &joint->d6.twistLimit.damping, 0.01f);
+						changed |= ImGui::DragFloat("Restitution Twist", &joint->d6.twistLimit.restitution, 0.01f);
+						changed |= ImGui::DragFloat("Bounce Threshold Twist", &joint->d6.twistLimit.bounceThreshold, 0.01f);
+						ImGui::Separator();
+
+						ImGui::Text("Swing");
+						changed |= ImGui::DragFloat("Swing Y", &joint->d6.swingLimit.yAngle, 0.01f);
+						changed |= ImGui::DragFloat("Swing Z", &joint->d6.swingLimit.zAngle, 0.01f);
+						changed |= ImGui::DragFloat("Stiffness Swing", &joint->d6.swingLimit.stiffness, 0.01f);
+						changed |= ImGui::DragFloat("Damping Swing", &joint->d6.swingLimit.damping, 0.01f);
+						changed |= ImGui::DragFloat("Restitution Swing", &joint->d6.swingLimit.restitution, 0.01f);
+						changed |= ImGui::DragFloat("Bounce Threshold Swing", &joint->d6.swingLimit.bounceThreshold, 0.01f);
+						ImGui::TreePop();
+					}
+
+					if (ImGui::TreeNode("Drives"))
+					{
+						changed |= ImGui::Checkbox("Drive Limits Are Forces", &joint->d6.driveLimitsAreForces);
+
+						auto drawDrive = [&](const char* label, Phy_D6JointDriveSettings& d)
+						{
+							if (ImGui::TreeNode(label))
+							{
+								changed |= ImGui::DragFloat("Stiffness", &d.stiffness, 0.01f);
+								changed |= ImGui::DragFloat("Damping", &d.damping, 0.01f);
+								changed |= ImGui::DragFloat("Force Limit", &d.forceLimit, 1.0f, 0.0f);
+								changed |= ImGui::Checkbox("Acceleration", &d.isAcceleration);
+								ImGui::TreePop();
+							}
+						};
+
+						drawDrive("Drive X", joint->d6.driveX);
+						drawDrive("Drive Y", joint->d6.driveY);
+						drawDrive("Drive Z", joint->d6.driveZ);
+						drawDrive("Drive Swing", joint->d6.driveSwing);
+						drawDrive("Drive Twist", joint->d6.driveTwist);
+						drawDrive("Drive Slerp", joint->d6.driveSlerp);
+						ImGui::TreePop();
+					}
+
+					if (ImGui::TreeNode("Drive Target"))
+					{
+						changed |= drawFrame("Drive Pose", joint->d6.drivePose);
+						changed |= ImGui::DragFloat3("Drive Linear Vel", &joint->d6.driveLinearVelocity.x, 0.01f);
+						changed |= ImGui::DragFloat3("Drive Angular Vel", &joint->d6.driveAngularVelocity.x, 0.01f);
+						ImGui::TreePop();
+					}
+					break;
+				}
+				}
+
+				if (changed) g_SceneDirty = true;
+			}
+		}
+	}
+}
